refactor(series): introduce plugin registry for series preparation
