+++ src/index.ts
@@ -1,3 +1,5 @@
+import './plugins';
+
 export {CustomShapeRenderer} from './utils';
 export {getFormattedValue} from '~core/utils';
 export {getDefaultTooltipHeaderFormat} from '~core/utils/tooltip';
