+++ src/plugins/bar-x/prepare.ts
@@ -1,39 +1,41 @@
 import type {ScaleOrdinal} from 'd3-scale';
 import get from 'lodash/get';
 
-import type {ChartSeriesOptions, HeatmapSeries, HeatmapSeriesData} from '../../types';
-import {DEFAULT_DATALABELS_STYLE} from '../constants';
-import {getUniqId} from '../utils';
+import {DEFAULT_DATALABELS_STYLE, seriesRangeSliderOptionsDefaults} from '~core/constants';
+import {DEFAULT_DATALABELS_PADDING} from '~core/series/constants';
+import type {PreparedBarXSeries, PreparedLegend, PreparedSeries} from '~core/series/types';
+import {getSeriesStackId, prepareLegendSymbol} from '~core/series/utils';
+import {getUniqId} from '~core/utils';
 
-import {DEFAULT_DATALABELS_PADDING} from './constants';
-import type {PreparedHeatmapSeries, PreparedLegend, PreparedSeries} from './types';
-import {prepareLegendSymbol} from './utils';
+import type {BarXSeries, BarXSeriesData, ChartSeriesOptions} from '../../types';
 
-type PrepareHeatmapSeriesArgs = {
+type PrepareBarXSeriesArgs = {
     colorScale: ScaleOrdinal<string, string>;
-    series: HeatmapSeries[];
+    series: BarXSeries[];
     legend: PreparedLegend;
     seriesOptions?: ChartSeriesOptions;
 };
 
-function prepareSeriesData(series: HeatmapSeries): HeatmapSeriesData[] {
+function prepareSeriesData(series: BarXSeries): BarXSeriesData[] {
     const nullMode = series.nullMode ?? 'skip';
     const data = series.data;
     switch (nullMode) {
         case 'zero':
-            return data.map((p) => ({...p, value: p.value ?? 0}));
+            return data.map((p) => ({...p, y: p.y ?? 0}));
         case 'skip':
         default:
             return data;
     }
 }
 
-export function prepareHeatmapSeries(args: PrepareHeatmapSeriesArgs): PreparedSeries[] {
+export function prepareBarXSeries(args: PrepareBarXSeriesArgs): PreparedSeries[] {
     const {colorScale, series: seriesList, seriesOptions, legend} = args;
 
-    return seriesList.map<PreparedHeatmapSeries>((series) => {
+    return seriesList.map<PreparedBarXSeries>((series) => {
         const name = series.name || '';
         const color = series.color || colorScale(name);
+        const dataLabelsInside =
+            series.stacking === 'percent' ? true : get(series, 'dataLabels.inside', false);
 
         return {
             type: series.type,
@@ -48,21 +50,23 @@ export function prepareHeatmapSeries(args: PrepareHeatmapSeriesArgs): PreparedSe
                 itemText: series.legend?.itemText ?? name,
             },
             data: prepareSeriesData(series),
+            stacking: series.stacking,
+            stackId: getSeriesStackId(series),
+            valueAxis: 'y',
             dataLabels: {
                 enabled: series.dataLabels?.enabled || false,
+                inside: dataLabelsInside,
                 style: Object.assign({}, DEFAULT_DATALABELS_STYLE, series.dataLabels?.style),
+                allowOverlap: series.dataLabels?.allowOverlap || false,
                 padding: get(series, 'dataLabels.padding', DEFAULT_DATALABELS_PADDING),
-                html: series.dataLabels?.html ?? false,
+                html: get(series, 'dataLabels.html', false),
                 format: series.dataLabels?.format,
             },
             cursor: get(series, 'cursor', null),
             yAxis: get(series, 'yAxis', 0),
+            borderRadius: get(series, 'borderRadius', get(seriesOptions, 'bar-x.borderRadius', 0)),
             tooltip: series.tooltip,
-            borderColor:
-                series.borderColor ??
-                seriesOptions?.heatmap?.borderColor ??
-                'var(--gcharts-shape-border-color)',
-            borderWidth: series.borderWidth ?? seriesOptions?.heatmap?.borderWidth ?? 0,
+            rangeSlider: Object.assign({}, seriesRangeSliderOptionsDefaults, series.rangeSlider),
         };
     }, []);
 }
