+++ src/plugins/radar/prepare.ts
@@ -2,18 +2,18 @@ import {scaleOrdinal} from 'd3-scale';
 import get from 'lodash/get';
 import merge from 'lodash/merge';
 
-import type {ChartSeriesOptions, RadarSeries} from '../../types';
-import {DEFAULT_DATALABELS_STYLE} from '../constants';
-import type {PointMarkerOptions} from '../types/chart/marker';
-import {getUniqId} from '../utils';
-
+import {DEFAULT_DATALABELS_STYLE} from '~core/constants';
 import {
     DEFAULT_DATALABELS_PADDING,
     DEFAULT_HALO_OPTIONS,
     DEFAULT_POINT_MARKER_OPTIONS,
-} from './constants';
-import type {PreparedLegend, PreparedRadarSeries} from './types';
-import {prepareLegendSymbol} from './utils';
+} from '~core/series/constants';
+import type {PreparedLegend, PreparedRadarSeries} from '~core/series/types';
+import {prepareLegendSymbol} from '~core/series/utils';
+import type {PointMarkerOptions} from '~core/types/chart/marker';
+import {getUniqId} from '~core/utils';
+
+import type {ChartSeriesOptions, RadarSeries} from '../../types';
 
 type PrepareRadarSeriesArgs = {
     series: RadarSeries[];
