+++ src/plugins/scatter/index.ts
@@ -1,0 +1,16 @@
+import type {SeriesPlugin} from '~core/series/plugin';
+
+import type {ScatterSeries} from '../../types';
+
+import {prepareScatterSeries} from './prepare';
+
+export const scatterPlugin: SeriesPlugin<ScatterSeries> = {
+    type: 'scatter',
+    prepareSeries: ({series, seriesOptions, legend, colorScale}) =>
+        prepareScatterSeries({
+            series: series as ScatterSeries[],
+            seriesOptions,
+            legend,
+            colorScale,
+        }),
+};
