+++ src/plugins/treemap/prepare.ts
@@ -1,13 +1,13 @@
 import type {ScaleOrdinal} from 'd3-scale';
 import get from 'lodash/get';
 
-import type {ChartSeriesOptions, TreemapSeries} from '../../types';
-import {DEFAULT_DATALABELS_STYLE, LayoutAlgorithm} from '../constants';
-import {getUniqId} from '../utils';
+import {DEFAULT_DATALABELS_STYLE, LayoutAlgorithm} from '~core/constants';
+import {DEFAULT_DATALABELS_PADDING} from '~core/series/constants';
+import type {PreparedLegend, PreparedTreemapSeries} from '~core/series/types';
+import {prepareLegendSymbol} from '~core/series/utils';
+import {getUniqId} from '~core/utils';
 
-import {DEFAULT_DATALABELS_PADDING} from './constants';
-import type {PreparedLegend, PreparedTreemapSeries} from './types';
-import {prepareLegendSymbol} from './utils';
+import type {ChartSeriesOptions, TreemapSeries} from '../../types';
 
 type PrepareTreemapSeriesArgs = {
     colorScale: ScaleOrdinal<string, string>;
